+++ packages/widget_test_harness/lib/widget_test_harness.dart
@@ -1,5 +1,4 @@
 export 'src/given_when_then.dart';
-export 'src/harness_setup/harness_setup.dart';
 export 'src/test_harness.dart';
 export 'src/ui_harness_helpers/given.dart';
 export 'src/ui_harness_helpers/when.dart';